% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStationList.R
\name{getStationList}
\alias{getStationList}
\title{getStationList}
\usage{
getStationList(hub, bounding.box = NA)
}
\arguments{
\item{hub}{The location of the API service. Default options include 'swmc', 'grand' and 'quinte'. See the README for adding additional hubs.}

\item{bounding.box}{(Optional) A comma separated string in the following format: min_x, min_y, max_x, max_y}
}
\value{
Dataframe containing station information to be passed to other functions (e.g. getTimeseriesList)
}
\description{
Returns a dataframe with station information. Can limit results with bounding box.
}
\examples{
\dontrun{
# All stations
station.dat <- getStationList()

# With bbox
station.dat <- getStationList(bounding.box = "-81.0269,43.4583,-79.0025,44.8189")
}
}
\author{
Ryan Whaley
}
