% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeseriesList.R
\name{getTimeseriesList}
\alias{getTimeseriesList}
\title{getTimeseriesList}
\usage{
getTimeseriesList(hub, station.id)
}
\arguments{
\item{hub}{The location of the API service. Default options include 'swmc', 'grand' and 'quinte'. See the README for adding additional hubs.}

\item{station_id}{A station identifier code. Can be found with getStationList()}
}
\value{
Dataframe containing a list of timeseries for the chosen station.
}
\description{
getTimeseriesList
}
\examples{
\dontrun{
 stn.ts <- getTimeseriesList("133535")
 }

}
\author{
Ryan Whaley
}
