% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ki_site_list.R
\name{ki_site_list}
\alias{ki_site_list}
\title{Get list of available sites}
\usage{
ki_site_list(hub, return_fields, datasource = 0)
}
\arguments{
\item{hub}{The KiWIS database you are querying. Either one of the defaults or a URL.
See \href{https://github.com/rywhale/kiwisR}{README}.}

\item{return_fields}{(Optional) KiWIS return fields as a list.}

\item{datasource}{(Optional) The data source to be used, defaults to 0.}
}
\value{
A tibble
}
\description{
Returns a tibble containing metadata available catchments.
}
\examples{
\dontrun{
ki_site_list(hub = 'swmc')
}

}
