% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ki_station_list.R
\name{ki_station_list}
\alias{ki_station_list}
\title{Get tibble containing station information.
Returns all available stations by default and allows for search terms.}
\usage{
ki_station_list(hub, search_term, bounding_box = NA)
}
\arguments{
\item{hub}{The KiWIS database you are querying. Default options are 'swmc', 'grand', 'quinte'. See README for more details.}

\item{search_term}{(Optional) A station name to search for. Supports the use of * as a wildcard.}

\item{bounding_box}{(Optional) A bounding box to search withhin for stations. Should be a vector or comma separated string
with the following format: (min_x, min_y, max_x, max_y).}
}
\value{
Tibble containing station id, name, latitude and longitude
}
\description{
Get tibble containing station information.
Returns all available stations by default and allows for search terms.
}
\examples{
ki_station_list(hub = 'swmc')
ki_station_list(hub = 'swmc', search_term = "A*")
ki_station_list(hub = 'swmc', search_term = "Lake Ontario at Toronto")

}
