% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ki_timeseries_list.R
\name{ki_timeseries_list}
\alias{ki_timeseries_list}
\title{Get list of available time series for station or
list of stations.}
\usage{
ki_timeseries_list(hub, station_id, ts_name)
}
\arguments{
\item{hub}{The KiWIS database you are querying. Default options are 'swmc', 'grand', 'quinte' and 'creditvalley'.
See README for more details.}

\item{station_id}{Either a single station id or a vector of station id. Can be string or numeric.
Station ids can be found using the ki_station_list function.}

\item{ts_name}{(Optional) A specific time series short name to search for. E.g. 'TAir.1.O'
Supports the use of "*" as a wildcard.}
}
\value{
A tibble containing all available time series for selected stations.
}
\description{
Get list of available time series for station or
list of stations.
}
\examples{
ki_timeseries_list(hub = 'swmc', station_id = "144659")
ki_timeseries_list(hub = 'swmc', station_id = c("144659", "144342"))

}
