% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchStations.R
\name{searchStations}
\alias{searchStations}
\title{searchStations()}
\usage{
searchStations(hub, search.trm)
}
\arguments{
\item{search.trm}{String to search station names for. Can use * as wildcard.}
}
\value{
Dataframe containing station information for stations with names matching search.trm
}
\description{
Returns a dataframe of stations with names matching the search string.
}
\examples{
\dontrun{
# All stations
station.A <- searchStations("A*")
head(station.A)
      station_name station_id station_latitude station_longitude
 1    ABBOTSFORD A     133535      49.03307961      -122.3666708
 2         ALERT A     132255      82.51666993       -62.2830803
 3      ALEXANDRIA     147654      45.31666964      -74.61666967

}
}
\author{
Ryan Whaley
}
