% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeseriesValues.R
\name{getTimeseriesValues}
\alias{getTimeseriesValues}
\title{getTimeseriesValues}
\usage{
getTimeseriesValues(hub, ts.id, from = NA, to = NA, time.stamp = TRUE)
}
\arguments{
\item{hub}{The location of the API service. Default options include 'swmc', 'grand' and 'quinte'. See the README for adding additional hubs.}

\item{from}{The start of the date range. Defaults to yesterday.}

\item{to}{The end of the date range. Defaults to today.}

\item{time.stamp}{Whether or not to return timestamp with date. Defaults to TRUE.}

\item{ts_id}{Timeseries identifier. Can be found using getTimeseriesList().}
}
\value{
Dataframe containing values for the specified date range and timeseries.
}
\description{
getTimeseriesValues
}
\examples{
\dontrun{
ts.dat <- getTimeseriesValues(ts_id = "948603042")
}

}
\author{
Ryan Whaley
}
