% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ki_group_list.R
\name{ki_group_list}
\alias{ki_group_list}
\title{Get list of available groups}
\usage{
ki_group_list(hub, datasource = 0)
}
\arguments{
\item{hub}{The KiWIS database you are querying. Either one of the defaults or a URL.
See \href{https://github.com/rywhale/kiwisR}{README}.}

\item{datasource}{(Optional) The data source to be used, defaults to 0.}
}
\value{
A tibble with three columns: group_id, group_name and group_type.
}
\description{
Returns a tibble containing metadata available groups. This can be used to
further filter down other queries like `ki_station_list`.
}
\examples{
\dontrun{
ki_group_list(hub = 'swmc')
}

}
