% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ki_group_list.R
\name{ki_group_list}
\alias{ki_group_list}
\title{Get list of available groups}
\usage{
ki_group_list(hub)
}
\arguments{
\item{hub}{The KiWIS database you are querying.
Default options are 'swmc', 'grand', 'quinte' and 'creditvalley'.
See README for more details.}
}
\value{
A tibble with three columns: group_id, group_name and group_type
}
\description{
Returns a tibble containing available groups. Allows users to query
ki_station_list for grouped stations.
}
\examples{
\dontrun{
ki_group_list(hub = 'swmc')
}
}
