% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ki_station_list.R
\name{ki_station_list}
\alias{ki_station_list}
\title{Get tibble containing station information.}
\usage{
ki_station_list(hub, search_term, bounding_box, group_id, return_fields)
}
\arguments{
\item{hub}{The KiWIS database you are querying. Default options are 'swmc', 'grand', 'quinte' and 'creditvalley'.
See README for more details.}

\item{search_term}{(Optional) A station name to search for. Supports the use of * as a wildcard. Case doesn't matter.}

\item{bounding_box}{(Optional) A bounding box to search withhin for stations. Should be a vector or comma separated string}

\item{group_id}{(Optional) A station group id (see ki_group_list).
with the following format: (min_x, min_y, max_x, max_y).}

\item{return_fields}{(Optional) Specific fields to return. Consult your KiWIS hub services documentation for available options.
Should be a comma separate string or a vector.}
}
\value{
Tibble containing station id, name, latitude and longitude
}
\description{
Returns all available stations by default and allows for search terms.
}
\examples{
\dontrun{
ki_station_list(hub = 'swmc')
ki_station_list(hub = 'swmc', search_term = "A*")
ki_station_list(hub = 'swmc', search_term = "Lake Ontario at Toronto")
ki_station_list(hub = 'swmc', group_id = '169270')
}
}
