% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ki_timeseries_values.R
\name{ki_timeseries_values}
\alias{ki_timeseries_values}
\title{Get values for time series id or list of time series ids.}
\usage{
ki_timeseries_values(hub, ts_id, start_date, end_date, return_fields)
}
\arguments{
\item{hub}{The KiWIS database you are querying. Either one of the defaults or a URL.
See \href{https://github.com/rywhale/kiwisR}{README}.}

\item{ts_id}{Either: a single time series id or a vector of time series ids.
Time series ids can be found using the ki_timeseries_list function}

\item{start_date}{A date string formatted "YYYY-MM-DD". Defaults to yesterday.}

\item{end_date}{A date string formatted "YYYY-MM-DD". Defaults to today.}

\item{return_fields}{(Optional) Specific fields to return. Consult your KiWIS hub services documentation for available options.
Should be a comma separate string or a vector.}
}
\value{
A tibble with following columns: Timestamp, Value, ts_name, Units, station_name

Tibble containing station metdata.
}
\description{
Returns time series values for given time series id and date range.
}
\examples{
\dontrun{
ki_timeseries_values(
  hub = "swmc",
  ts_id = "1125831042",
  start_date = "2015-12-01",
  end_date = "2018-01-01"
)
}

}
